requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
python-multipart==0.0.5
yara-python==4.1.3 
pydantic==1.9.0 
